requirements.txt
dj-database-url==0.5.0
Django==1.11
psycopg2==2.7.4
psycopg2-binary==2.7.4
python-decouple==3.1
pytz==2018.3
djangorestframework==3.7.7
Markdown==2.6.11
django-filter==1.1.0
